Base Imathx include headers on current Imath implementation
The ImathExc header has been removed in the current Imath implementation,
which is already installed on MacOS for example. The header is still
in use in OSL master, which makes OSL installations unusable on at least some
Macs, and likely other systems as well.

This commit updates the Imathx headers, to throw std::invalid_argument,
and std::domain_error instead of SingMatrixExc, and NullVecExc respectively,
mirroring the current Imath headers.